requirements.txt
cffi==1.16.0
cryptography==42.0.8
gitdb==4.0.11
GitPython==3.1.43
numpy==1.26.4
oracledb==2.1.1
pandas==2.2.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2024.1
six==1.16.0
smmap==5.0.1
tzdata==2024.1
